Second Stage
A new custom user model is created, as well as a second view, the
rent view.

User authentication is done using tokens. In order to post, patch
and delete from the Movies view, the user must be a staff member.
Tokens are passed through the `curl` command, and are created
for an already existing user by running
`./manage.py drf_create_token <username>` .

Currently, if an authenticated user sends a GET request to
/api/rent/<movie id>, this movie id is added to his movies_rented
list.

+++ movies_store/base_app/views.py
@@ -2,8 +2,8 @@ from django.shortcuts import get_object_or_404
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from .models import Movie
-from .serializers import MovieSerializer
+from .models import Movie, User
+from .serializers import MovieSerializer, UserSerializer
 
 # Create your views here.
 class MoviesView(APIView):
@@ -23,21 +23,48 @@ class MoviesView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        serializer = MovieSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        if request.user.is_staff:
+            serializer = MovieSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({"status": "401 Unauthorized",
+                            "data": "You are not authorized to perform this action"},
+                            status=status.HTTP_401_UNAUTHORIZED)
 
     def patch(self, request, id=None):
-        movie = Movie.objects.get(id=id)
-        serializer = MovieSerializer(movie, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_staff:
+            movie = Movie.objects.get(id=id)
+            serializer = MovieSerializer(movie, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({"status": "401 Unauthorized",
+                            "data": "You are not authorized to perform this action"},
+                            status=status.HTTP_401_UNAUTHORIZED)
 
     def delete(self, request, id=None):
-        movie = get_object_or_404(Movie, id=id)
-        movie.delete()
-        return Response({"status": "success", "data": "Movie Deleted"})
+        if request.user.is_staff:
+            movie = get_object_or_404(Movie, id=id)
+            movie.delete()
+            return Response({"status": "success", "data": "Movie Deleted"})
+        else:
+            return Response({"status": "401 Unauthorized",
+                            "data": "You are not authorized to perform this action"},
+                            status=status.HTTP_401_UNAUTHORIZED)
+
+
+class RentView(APIView):
+
+    def get(self, request, id):
+        if request.user.is_authenticated:
+            movie = Movie.objects.get(id=id)
+            this_user = User.objects.get(id=request.user.id)
+            this_user.movies_rented.add(movie)
+            serializer = UserSerializer(this_user)
+            return Response(serializer.data, status=status.HTTP_200_OK)
